+++ offside/lib/MyPage/profile.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dropdown_textfield/dropdown_textfield.dart';
-import 'package:offside/user_view_model.dart';
+import 'package:offside/data/view/user_view_model.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class Edit extends ConsumerStatefulWidget {
