+++ offside/lib/data/model/match_model.dart
@@ -1,0 +1,12 @@
+class MatchModel {
+  MatchModel(this.data, this.score2, this.team1, this.team2, this.location,
+      this.time, this.score1);
+
+  String? data;
+  String? score2;
+  String? team1;
+  String? team2;
+  String? location;
+  String? time;
+  String? score1;
+}
