+++ offside/lib/data/model/team_transfer.dart
@@ -1,0 +1,85 @@
+import 'package:flutter/widgets.dart';
+
+List<String> team1 = <String>[
+  "강원 FC",
+  "광주 FC",
+  "대구 FC",
+  "대전\n하나 시티즌",
+  "FC 서울",
+  "수원 삼성\n블루윙즈",
+  "수원 FC",
+  "울산 현대",
+  "인천\n유나이티드",
+  "전북 현대\n모터스",
+  "제주\n유나이티드",
+  "포항\n스틸러스"
+];
+List<String> team2 = <String>[
+  "경남 FC",
+  "김천 상무",
+  "김포 FC",
+  "부산\n아이파크",
+  "부천 FC\n1995",
+  "서울 이랜드",
+  "성남 FC",
+  "안산\n그리너스",
+  "FC 안양",
+  "전남\n드래곤즈",
+  "충남\n아산 FC",
+  "충북\n청주 FC",
+  "천안\n시티 FC"
+];
+Map<String, String> transferName = {
+  '강원 FC': '강원',
+  '광주 FC': '광주',
+  '대구 FC': '대구',
+  '대전 하나 시티즌': '대전',
+  'FC 서울': '서울',
+  '수원 삼성 블루윙즈': '수원',
+  '수원 FC': '수원FC',
+  '울산 현대': '울산',
+  '인천 유나이티드': '인천',
+  '전북 현대 모터스': '전북',
+  '제주 유나이티드': '제주',
+  '포항 스틸러스': '포항',
+  '경남 FC': '경남',
+  '김천 상무 FC': '김천',
+  '김포 FC': '김포',
+  '부산 아이파크': '부산',
+  '부천 FC 1995': '부천',
+  '서울 이랜드 FC': '서울E',
+  '성남 FC': '성남',
+  '안산 그리너스 FC': '안산',
+  'FC 안양': '안양',
+  '전남 드래곤즈': '전남',
+  '충남 아산 FC': '아산',
+  '충북 청주 FC': '청주',
+  '천안 시티 FC': '천안'
+};
+Map<String, dynamic> teamImg = {
+  '강원 FC': Image.asset('images/K1_png/Gangwon.png'),
+  '광주 FC': Image.asset('images/K1_png/GwangJu.png'),
+  '대구 FC': Image.asset('images/K1_png/Daegu.png'),
+  '대전 하나 시티즌': Image.asset('images/K1_png/Daejeon.png'),
+  'FC 서울': Image.asset('images/K1_png/Seoul.png'),
+  '수원 삼성 블루윙즈': Image.asset('images/K1_png/suwon.png'),
+  '수원 FC': Image.asset('images/K1_png/suwonFC.png'),
+  '울산 현대': Image.asset('images/K1_png/ulsan.png'),
+  '인천 유나이티드': Image.asset('images/K1_png/incheon.png'),
+  '전북 현대 모터스': Image.asset('images/K1_png/jeonbuk.png'),
+  '제주 유나이티드': Image.asset('images/K1_png/jeju.png'),
+  '포항 스틸러스': Image.asset('images/K1_png/pohang.png'),
+  '경남 FC': Image.asset('images/K2_png/kn.png'),
+  '김천 상무 FC': Image.asset('images/K2_png/kimcheon.png'),
+  '김포 FC': Image.asset('images/K2_png/kimpo.png'),
+  '부산 아이파크': Image.asset('images/K2_png/busan.png'),
+  '부천 FC 1995': Image.asset('images/K2_png/bc.jpg'),
+  '서울 이랜드 FC': Image.asset('images/K2_png/seoulE.png'),
+  '성남 FC': Image.asset('images/K2_png/seongnam.png'),
+  '안산 그리너스 FC': Image.asset('images/K2_png/ansan.png'),
+  'FC 안양': Image.asset('images/K2_png/anyang.png'),
+  '전남 드래곤즈': Image.asset('images/K2_png/jeonnam.png'),
+  '충남 아산 FC': Image.asset('images/K2_png/asan.png'),
+  '충북 청주 FC': Image.asset('images/K2_png/chungju.png'),
+  '천안 시티 FC': Image.asset('images/K2_png/cheonan.png')
+};
