+++ offside/lib/data/model/tour_model.dart
@@ -1,0 +1,10 @@
+import 'dart:convert';
+
+class TourModel {
+  TourModel(this.addr, this.contentId, this.typeId, this.img, this.title);
+  String? addr;
+  String? contentId;
+  String? typeId;
+  String? img;
+  String? title;
+}
