+++ offside/lib/main.dart
@@ -1,11 +1,13 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:offside/data/datasource/tour_data_source.dart';
+import 'package:offside/data/view/match_view_model.dart';
+import 'package:offside/data/view/tour_view_model.dart';
+import 'package:offside/data/view/user_view_model.dart';
 import 'package:offside/firebase_options.dart';
 import 'package:offside/login/login.dart';
 import 'package:kakao_flutter_sdk_common/kakao_flutter_sdk_common.dart';
-import 'package:offside/user_view_model.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:offside/MainPage/main_page.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -30,6 +32,9 @@ class Offside extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final autoLogin = ref.read(userViewModelProvider).autoSignIn();
+    final matchData = ref.read(matchViewModelProvider).getAllMatches();
+    final tmp = ref.read(tourViewModelProvider).getTourData();
+
     return MaterialApp(
         title: 'Offside',
         initialRoute: '/',
@@ -39,6 +44,7 @@ class Offside extends ConsumerWidget {
               seedColor: const Color.fromARGB(255, 255, 255, 255)),
           useMaterial3: true,
         ),
-        home: autoLogin == true ? const MainPage() : LoginPage());
+        // home: autoLogin == true ? const MainPage() : LoginPage());
+        home: LoginPage());
   }
 }
