+++ offside/pubspec.yaml
@@ -37,7 +37,7 @@ dependencies:
   firebase_core: ^2.14.0
   firebase_auth: ^4.6.3
   kakao_flutter_sdk_user: ^1.4.3
-  http: ^1.0.0
+  http: ^1.1.0
   dropdown_textfield: ^1.0.8
   carousel_slider: ^4.2.1
   flutter_svg: ^1.0.1
